move node task transforms to task_transforms.py; remove unnecessary checks
